first iteration of adding Liquibase

+++ src/main/java/com/jpr/maintenance/database/repository/TaskDetailsRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TaskDetailsRepository extends CrudRepository<TaskDetailsEntity, Integer> {
+public interface TaskDetailsRepository extends CrudRepository<TaskDetailsEntity, Long> {
 }
